+++ models/user.model.js
@@ -3,7 +3,7 @@ const bcrypt = require('bcrypt');
 module.exports = (sequelize, DataTypes) => {
     const User = sequelize.define('user',
     {
-        id: {type: DataTypes.STRING, primaryKey: true, allowNull: false},
+        userId: {type: DataTypes.STRING, primaryKey: true, allowNull: false},
         password: {type: DataTypes.STRING, allowNull: false}
     },
     {
